requirements.txt
absl-py
aiohttp
async-timeout
attrs
blinker==1.4
brotlipy
cachetools
certifi
cffi
chardet
click
cryptography
dataclasses==0.8
idna==2.10
importlib-metadata
joblib
markdown
numpy
oauthlib
pillow
pip
point_cloud_utils
protobuf
pyasn1
pyasn1-modules
pycparser==2.20
pyjwt
pyopenssl
pysocks
pytorch
pyu2f
requests
requests-oauthlib
rsa
scikit-learn
scipy
setuptools
six
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
threadpoolctl
torchvision
tqdm
geomloss
typing-extensions
typing_extensions
urllib3
yarl
zipp
